+++ CRM_be/src/main/java/org/example/crm_be/module/customers/infrastructure/persistence/CustomerJpaRepository.java
@@ -5,5 +5,5 @@ import org.example.crm_be.module.customers.domain.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerJpaRepository
-        extends JpaRepository<Customer, Long> {
+        extends JpaRepository<Customer, Integer> {
 }
